read_opj: fix comparison between signed and unsigned integers

+++ src/read_opj.cpp
@@ -39,14 +39,14 @@ static DataFrame import_spreadsheet(const Origin::SpreadSheet & osp, const char 
 			)
 		){
 			NumericVector ncol(osp.maxRows, NA_REAL);
-			for (int row = 0; row < std::min(ocol.data.size(), (size_t)osp.maxRows); row++) {
+			for (size_t row = 0; row < std::min(ocol.data.size(), (size_t)osp.maxRows); row++) {
 				ncol[row] = ocol.data[row].as_double();
 				if (ncol[row] == _ONAN) ncol[row] = R_NaN;
 			}
 			rsp[c] = ncol;
 		} else {
 			StringVector ccol(osp.maxRows, NA_STRING);
-			for (int row = 0; row < std::min(ocol.data.size(), (size_t)osp.maxRows); row++) {
+			for (size_t row = 0; row < std::min(ocol.data.size(), (size_t)osp.maxRows); row++) {
 				const Origin::variant & v = ocol.data[row];
 				if (v.type() == Origin::variant::V_DOUBLE) {
 					if (v.as_double() != _ONAN) ccol[row] = std::to_string(v.as_double()); // yuck
